+++ src/consumer/component/sources/data-to-persist.js
@@ -30,6 +30,7 @@ export default class DataToPersist {
         return;
       }
     }
+    this._throwForDirectoryCollision(file);
     this.files.push(file);
   }
   addManyFiles(files: AbstractVinyl[] = []) {
@@ -140,6 +141,26 @@ export default class DataToPersist {
   _assertRelative(pathToCheck: string) {
     if (path.isAbsolute(pathToCheck)) {
       throw new Error(`DataToPersist expects ${pathToCheck} to be relative, but found it absolute`);
+    }
+  }
+  /**
+   * prevent adding a file which later on will cause an error "EEXIST: file already exists, mkdir {dirname}".
+   * this happens one a file is a directory name of the other file.
+   * e.g. adding these two files, will cause the error above: "bar/foo" and "bar"
+   *
+   * to check for this possibility, we need to consider two scenarios:
+   * 1) "bar/foo" is there and now adding "bar" => check whether one of the files starts with "bar/"
+   * 2) "bar" is there and now adding "bar/foo" => check whether this file "bar/" starts with one of the files
+   */
+  _throwForDirectoryCollision(file: AbstractVinyl) {
+    const directoryCollision = this.files.find(
+      f => f.path.startsWith(`${file.path}${path.sep}`) || `${file.path}${path.sep}`.startsWith(f.path)
+    );
+    if (directoryCollision) {
+      throw new Error(`unable to add the file "${file.path}", because another file "${
+        directoryCollision.path
+      }" is going to be written.
+one of them is a directory of the other one, and is not possible to have them both`);
     }
   }
 }